+++ packages/a2ui_core/test/listenable/smoke_test.dart
@@ -1,0 +1,88 @@
+// Copyright 2025 The Flutter Authors.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'package:a2ui_core/src/listenable/notifiers.dart';
+import 'package:a2ui_core/src/listenable/primitives.dart';
+import 'package:test/test.dart';
+
+// ignore: unused_element, tests that ValueNotifier can be implemented.
+class _ValueNotifierImplementation<T> implements GenUiValueNotifier<T> {
+  @override
+  void addListener(VoidCallback listener) {}
+
+  @override
+  void removeListener(VoidCallback listener) {}
+
+  @override
+  T get value => throw UnimplementedError();
+
+  @override
+  void dispose() {}
+
+  @override
+  set value(T newValue) {}
+
+  @override
+  bool get hasListeners => throw UnimplementedError();
+
+  @override
+  void notifyListeners() {}
+}
+
+// ignore: unused_element, tests that ValueNotifier can be extended.
+class _ValueNotifierExtension<T> extends GenUiValueNotifier<T> {
+  _ValueNotifierExtension(super.value);
+}
+
+// ignore: unused_element, tests that ChangeNotifier can be implemented.
+class _ChangeNotifierImplementation implements GenUiChangeNotifier {
+  @override
+  void addListener(VoidCallback listener) {}
+
+  @override
+  void removeListener(VoidCallback listener) {}
+
+  @override
+  void dispose() {}
+
+  @override
+  bool get hasListeners => throw UnimplementedError();
+
+  @override
+  void notifyListeners() {}
+}
+
+// ignore: unused_element, tests that ChangeNotifier can be extended.
+class _ChangeNotifierExtention extends GenUiChangeNotifier {}
+
+void main() {
+  test('ValueNotifier basic functionality is working', () {
+    final GenUiValueNotifier<int> notifier = GenUiValueNotifier(1);
+    addTearDown(notifier.dispose);
+    var count = 0;
+    notifier.addListener(() => count++);
+
+    expect(notifier, isA<GenUiValueListenable<int>>());
+    expect(notifier.value, 1);
+    expect(count, 0);
+
+    notifier.value = 2;
+    expect(notifier.value, 2);
+    expect(count, 1);
+  });
+
+  test('ChangeNotifier basic functionality is working', () {
+    final notifier = GenUiChangeNotifier();
+    addTearDown(notifier.dispose);
+    var count = 0;
+    notifier.addListener(() => count++);
+
+    expect(notifier, isA<GenUiListenable>());
+    expect(count, 0);
+
+    // ignore: invalid_use_of_protected_member
+    notifier.notifyListeners();
+    expect(count, 1);
+  });
+}